Goodbye channel fns, hello component-meta (#27)
The channel fns were confusing and we couldn't find a good use for them. This removes them.

A `::ds/component-meta` signal handler key is introduced. Its value is an atom, and you can use it to convey state across the entire lifecycle of a signal application. The final value of the atom is stored in the system map under `::component-meta`. This example should illustrate:


```clojure
(deftest component-meta-test
  (let [system {::ds/defs {:group {:component #::ds{:pre-start  (fn [{:keys [::ds/component-meta]}]
                                                                  (reset! component-meta [::ds/pre-start]))
                                                    :start      (fn [{:keys [::ds/component-meta]}]
                                                                  (swap! component-meta conj ::ds/start))
                                                    :post-start (fn [{:keys [::ds/component-meta]}]
                                                                  (swap! component-meta conj ::ds/post-start))}}}}]
    (is (= [::ds/pre-start ::ds/start ::ds/post-start]
           (get-in (ds/start system) [::ds/component-meta :group :component])))))
```

donut.error is now used for schema validation, including in the `donut.system.validation` namespace

pre-start and post-start stage can now be defined as maps of `{keyword function}` to allow multiple such stage handlers